Move registry search of install folder to EliteFiles
